requirements.txt
africastalking==1.2.8
Authlib==1.3.1
Automat==0.8.0
backcall==0.2.0
bcrypt==4.1.3
flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-CKEditor==1.0.0
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-Reuploaded==1.4.0
Flask-Reuploads==1.0.1.dev0
flask-sqlalchemy==3.1.1
Flask-Testing==0.8.1
Flask-Uploads==0.2.1
git-filter-repo==2.45.0
google-ai-generativelanguage==0.1.0
google-api-core==2.19.1
google-auth==2.32.0
google-auth-oauthlib==1.2.1
google-generativeai==0.1.0rc1
googleapis-common-protos==1.63.2
gunicorn==22.0.0
mysql-connector-python==9.0.0
pytest==8.3.3
python-dotenv==1.0.1
regex==2024.7.24
requests==2.22.0
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
SQLAlchemy==2.0.23
